ini tugas uas ke-4

+++ routers/users.js
@@ -1,18 +1,21 @@
 const mysql = require('mysql')
 const express = require('express')
+const auth = require('../middleware/auth.js')
 const router = express.Router();
+const bcrypt = require('bcrypt')
+const saltRounds = 10
 const con = mysql.createConnection({
     host: "localhost",
     user: "root",
     password: "",
-    database: "tugas-web"
+    database: "pasarsenin"
 })
 
 con.connect(function (err) {
     if (err) throw err
 })
 
-router.get('/user', function (req, res) {
+router.get('/', auth, function (req, res) {
     const sql = "SELECT * FROM user"
     con.query(sql, function (err, result) {
         if (err) throw err
@@ -21,22 +24,40 @@ router.get('/user', function (req, res) {
     })
 })
 
-router.post('/user', function (req, res) {
-    console.log(req.body)
-    const sql = "INSERT INTO user(username, password) VALUES ('" + req.body.username + "' , '" + req.body.password + "')"
-    con.query(sql, function (err) {
-        if (err) throw err
-        console.log("1 record inserted")
+router.post('/', function (req, res, next) {
+    const sql = "SELECT * FROM user"
+    con.query(sql, function (err, result) {
+        if (result.length > 0) {
+            auth(req, res, next)
+        } else {
+            next()
+        }
     })
-    res.end()
+}, function (req, res) {
+    bcrypt.genSalt(saltRounds, function (err, salt) {
+        bcrypt.hash(req.body.password, salt, function (err, hash) {
+            const sql = "INSERT INTO user(username, password) VALUES ('" + req.body.username + "' , '" + hash + "')"
+            con.query(sql, function (err) {
+                if (err) throw err
+                console.log("1 record inserted")
+                res.end()
+            })
+        })
+    })
 })
 
-router.delete('/user/:id', function (req, res) {
-    console.log(req.body)
-    const sql = "DELETE FROM user WHERE id = " + req.params.id
-    con.query(sql, function (err) {
-        if (err) throw err
-        console.log("1 record deleted")
+router.delete('/:id', auth, function (req, res) {
+    const check = "SELECT * FROM user"
+    con.query(check, function (err, result) {
+        if (result.length == 1)
+            return
+        else {
+            const sql = "DELETE FROM user WHERE id = " + req.params.id
+            con.query(sql, function (err) {
+                if (err) throw err
+                console.log("1 record deleted")
+            })
+        }
     })
     res.end()
 })
